requirements.txt
python-dotenv
pydantic
pdf2image
pillow
pytesseract
google-genai
requests
pandas
rich
streamlit
